Create ViewModels and link them to Views via Autofac
On the one hand, this simplifies a lot of the glue code that, before,
was basically just simulating Autofac. On the other hand, this adds a
lot of complexity.

It's unclear if this was the right decision, but I think having
dependency injection for ViewModels will prove to be very useful as the
GUI starts attaining functionality.

Signed-off-by: Joshua Brot <jbrot@umich.edu>